requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.3
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
cryptography==42.0.8
Deprecated==1.2.18
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.104.1
frozenlist==1.4.1
googleapis-common-protos==1.69.2
greenlet==3.0.1
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huey==2.5.0
humanfriendly==10.0
idna==3.4
importlib_metadata==8.6.1
iniconfig==2.0.0
Levenshtein==0.27.1
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-sqlalchemy==0.52b1
opentelemetry-instrumentation-system-metrics==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
packaging==24.0
passlib==1.7.4
pluggy==1.4.0
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
RapidFuzz==3.12.2
redis==5.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.17.2
yarl==1.9.4
zipp==3.21.0
